JENKINS-38670 move tests, bugfix template not showing up
